Patchnotes v1.00
Added Time.java, a class that simulates a clock in 24 hour time

Time() constructor takes 3 parameters, in the order of int int int. It initializes the instance variables of Time.java. Returns nothing

tick() method takes no parameters and is a void. It increments seconds by 1, and adjusts the other instance variables in a manner appropriate to a clock

add() method takes an instance of the Time object, and is a void. It increments all instance variables by the instance variables of the instance in the parameter

toString() method takes no parameters. It returns a String containing all the instance variables in HH:MM:SS format, as would an analog clock using 24 hour time